refactor: move TopicSubscribe.Item to TopicItem
